Pipe-friendly wrappers for external fitters

These functions provide pipe-friendly wrappers around model fitters provided by several external packages. The functions require the corresponding packages to be installed, if the required package is missing the functions warns with directions for how to install it.

zlm_robust() wraps estimatr::lm_robust(), which fits a linear model with a variety of options for estimating robust standard errors.

zpolr() wraps MASS::polr(), which fits an ordered logistic response for multi-value ordinal variables, using a proportional odds logistic regression.

zplsr() wraps pls::plsr(), which performs a partial least squares regression.

Examples

if (requireNamespace("estimatr") && getRversion() >= "4.1.0")
  zlm_robust(cars, dist ~ speed) |> summary()
#> Warning: 'zlm_robust()' requires the 'estimatr' package.
#>   Install with install.packages("estimatr") and restart R.
#>    Length     Class      Mode 
#>         1 character character 

if (requireNamespace("MASS") && getRversion() >= "4.1.0")
  zpolr(mtcars, ordered(gear) ~ mpg + hp) |> summary()
#> 
#> Re-fitting to get Hessian
#> Call:
#> polr(formula = ordered(gear) ~ mpg + hp, data = mtcars)
#> 
#> Coefficients:
#>       Value Std. Error t value
#> mpg 0.37279   0.123027   3.030
#> hp  0.02002   0.009634   2.078
#> 
#> Intercepts:
#>     Value   Std. Error t value
#> 3|4 10.1578  3.6605     2.7750
#> 4|5 12.7982  4.0427     3.1657
#> 
#> Residual Deviance: 51.16071 
#> AIC: 59.16071 

if (requireNamespace("pls") && getRversion() >= "4.1.0")
  zplsr(cars, dist ~ speed) |> summary()
#> Warning: 'zplsr()' requires the 'pls' package.
#>   Install with install.packages("pls") and restart R.
#>    Length     Class      Mode 
#>         1 character character